Save Load Update
First attempt at trying to have data persist through battle scenes but didn't get pretty far. Trying again later.

+++ Darkness/Assets/Scripts/SaveLoadSystem/SavedInfo.cs
@@ -5,11 +5,11 @@ using UnityEngine;
 [System.Serializable]
 public class SavedInfo
 {
-    public int health;
+    public float health;
     public Vector3 position;
 
     public SavedInfo(){
-        this.health = 30;
+        this.health = 30.0f;
         position = Vector3.zero;
     }
 }
